Completyed thisd project with user relationship.

+++ app/Http/Controllers/ListingController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use auth;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
@@ -21,7 +22,7 @@ class ListingController extends Controller
     //Show single listing
     public function show(Listing $listing) {
         return view('listings.show', [
-            'listing' => $listing
+            'listing' => $listing,  'listings' => auth()->user()->listings()->get()
         ]);
     }
 
@@ -47,6 +48,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
 
@@ -60,7 +63,12 @@ class ListingController extends Controller
 
     // Update listing data
     public function update(Request $request, Listing $listing) {
-        // dd($request->file('logo')); 
+        
+        //Make sure loged in user is owner
+        if($listing->user_id == auth()->id()) {
+            abort(403, 'Unauthorized');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'], 
@@ -82,9 +90,18 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing) {
+        if($listing->user_id == auth()->id()) {
+            abort(403, 'Unauthorized');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+    
+    //Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 
 }
